+++ src/main/java/insane96mcp/iguanatweaksreborn/module/items/StackSizes.java
@@ -31,11 +31,8 @@ public class StackSizes extends Feature {
     public static final TagKey<Item> NO_STACK_SIZE_CHANGES = ISOItemTagsProvider.create("no_stack_size_changes");
 
     @Config
-    @Label(name = "Food Stack Reduction", description = "Food stack sizes will be reduced based off their hunger restored and saturation multiplier. See 'Food Stack Reduction Formula' for the formula")
-    public static Boolean foodStackReduction = true;
-    @Config
-    @Label(name = "Food Stack Reduction Formula", description = "The formula to calculate the stack size of a food item. Variables as hunger, saturation_modifier, effectiveness as numbers and fast_food as boolean can be used. This is evaluated with EvalEx https://ezylang.github.io/EvalEx/concepts/parsing_evaluation.html.")
-    public static String foodStackReductionFormula = "ROUND((1 - (effectiveness - 1) / 23) * 64 * 0.5, 0)";
+    @Label(name = "Food Stack Reduction Formula", description = "The formula to calculate the stack size of a food item. Variables as hunger, saturation_modifier, effectiveness as numbers and fast_food as boolean can be used. Set to empty to disable this. This is evaluated with EvalEx https://ezylang.github.io/EvalEx/concepts/parsing_evaluation.html.")
+    public static String foodStackReductionFormula = "ROUND((1 - hunger / 9) * 64 * 0.5, 0)";
     @Config(min = 0.01d, max = 64d)
     @Label(name = "Item Stack Multiplier", description = "Items max stack sizes (excluding blocks) will be multiplied by this value. Foods will be overridden by 'Food Stack Reduction' or 'Food Stack Multiplier' if are active. Setting to 1 will disable this feature.")
     public static Double itemStackMultiplier = 1d;
@@ -112,7 +109,7 @@ public class StackSizes extends Feature {
     //Food
     @SuppressWarnings("deprecation")
     public static void processFoodStackSizes(boolean isClientSide) {
-        if (!foodStackReduction)
+        if (foodStackReductionFormula.isEmpty())
             return;
 
         for (Map.Entry<Item, Integer> entry : originalStackSizes.entrySet()) {
@@ -134,9 +131,9 @@ public class StackSizes extends Feature {
             return;
 
         if (event.getPlayer() == null)
-            event.getPlayerList().getPlayers().forEach(player -> StackSizesSync.sync(foodStackReduction, foodStackReductionFormula, itemStackMultiplier, blockStackMultiplier, player));
+            event.getPlayerList().getPlayers().forEach(player -> StackSizesSync.sync(foodStackReductionFormula, itemStackMultiplier, blockStackMultiplier, player));
         else
-            StackSizesSync.sync(foodStackReduction, foodStackReductionFormula, itemStackMultiplier, blockStackMultiplier, event.getPlayer());
+            StackSizesSync.sync(foodStackReductionFormula, itemStackMultiplier, blockStackMultiplier, event.getPlayer());
     }
 
     @SubscribeEvent(priority = EventPriority.HIGH)